More profile tests, more to come

+++ route/profile-router.js
@@ -12,7 +12,10 @@ const profileRouter = module.exports = Router();
 profileRouter.post('/api/hospital/:hospitalID/profile', bearerAuth, jsonParser, function(req, res, next){
   debug('hit POST route /api/hospital/:hospitalID/profile');
   new Profile(req.body).save()
-  .then( profile => res.json(profile))
+  .then( profile => {
+    console.log('HEHREHRHEHREHRHEHRE  ', profile);
+    res.json(profile);
+  })
   .catch(next);
 });
 
@@ -33,6 +36,7 @@ profileRouter.delete('/api/hospital/:hospitalID/profile/:profileID', bearerAuth,
   debug('Hit DELETE /api/hospital/:hospitalID/profile/:profileID');
   Profile.findById(req.params.profileID)
   .then(profile => {
+    if(profile.hospitalID.toString() !== req.params.hospitalID) return Promise.reject(createError(404, 'Profile not found.'));
     if(profile.userID.toString() === req.user._id.toString()) {
       Profile.findByIdAndRemove(req.params.profileID)
       .then(() => res.sendStatus(204))
@@ -48,6 +52,7 @@ profileRouter.put('/api/hospital/:hospitalID/profile/:profileID', bearerAuth, js
   debug('Hit PUT /api/hospital/:hospitalID/profile/:profileID');
   Profile.findById(req.params.profileID)
   .then(profile => {
+    if(profile.hospitalID.toString() !== req.params.hospitalID) return Promise.reject(createError(404, 'Profile not found.'));
     if(profile.userID.toString() === req.user._id.toString()) {
       Profile.findByIdAndUpdate(req.params.profileID, req.body, {new:true})
       .then(profile => res.json(profile))
